requirements.txt
streamlit
pandas
tabulate
langdetect
python-dotenv
google-generativeai
